api/requirements.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
ffmpeg-python==0.2.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
prompt-toolkit==3.0.16
pydub==0.24.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
youtube-dl==2021.2.10
zipp==3.4.0
